Copy test HTML file and add data source option
Instead of keeping an index.html in src/, store it in test/ so that
the src/ dir contains only chart builder source files (the HTML file is
only used for demo purposes and isn't a requirement).

In order to make testing easier, allow the browser to override the data
files' URL root so that sample data files can be loaded from localhost
(see test-runner.js).